Update settings migrators
