+++ src/settings/versions/1.1.0/index.ts
@@ -1,4 +1,4 @@
-import { ChatCompletionsModel, ChatHistory, CompletionsModel } from './openai';
+import { ChatCompletionsModel, ChatHistory, CompletionsModel } from './api/openai';
 
 export interface MarkpilotSettings1_1_0 {
   apiKey: string | undefined;
