+++ src/settings/versions/1.2.0/api/providers/index.ts
@@ -1,0 +1,10 @@
+export type Provider = (typeof PROVIDERS)[number];
+
+export const PROVIDERS = ['openai', 'openrouter', 'ollama'] as const;
+export const PROVIDERS_NAMES: Record<Provider, string> = {
+  openai: 'OpenAI',
+  openrouter: 'OpenRouter',
+  ollama: 'Ollama',
+};
+
+export const DEFAULT_PROVIDER = 'openai' as Provider;
